fix: complete terminal lifecycle extraction without losing close behavior
